+++ src/LeadII_V.cpp
@@ -1,7 +1,7 @@
 
 #include "LeadII_V.h"
 
-leadII_V::leadII_V(leads_name type, info_for_new_peak* ptr_info_new_peak, int N_leads, vector<double>* ptr_signal, one_lead* ptr_leadV)
+leadII_V::leadII_V(leads_name type, info_for_new_peak* ptr_info_new_peak, int N_leads, vector<float>* ptr_signal, one_lead* ptr_leadV)
 {
 	this->N_leads = N_leads;
 	type_of_lead = type;
