+++ src/check_last_four_peaks.cpp
@@ -38,7 +38,7 @@ bool leadII_V::check_last_four_peaks( vector<pair<int,pat_name > > & peaks, cons
 
         if (peak1 - peak3 < 1.1 * average_interval && peak1 - peak3 > 0.9 * average_interval &&
             //(peaks.end() - 2)->second != V_b && (peaks.end() - 2)->second != N_b) {
-            (double) (abs((peak2 - peak3)/(peak1 - peak2))) > 0.9 ){
+            (float) (abs((peak2 - peak3)/(peak1 - peak2))) > 0.9 ){
             peaks.erase(peaks.end() - 2);
             result = true;
         }
@@ -64,7 +64,7 @@ bool leadII_V::check_SV_A_extrasustole( vector<pair<int,pat_name > > & peaks, co
         if ((peaks.end() - 2)->second == N_b && (peaks.end() - 1)->second == N_b &&
             (peaks.end() - 3)->second == N_b)
             while (peak2 != 0 && peak3 != 0 && abs(peak2 - peak3) < average_R && (peak1 - peak2) > 0.95 * average_R
-                   && static_cast<double>(peak2 - peak3) / (peak1 - peak2) < 0.75 &&
+                   && static_cast<float>(peak2 - peak3) / (peak1 - peak2) < 0.75 &&
                    peak1 - peak3 < (2 + count_cur / 2.0) * 1.05 * average_R) {
                 count_cur++;
                 result = true;
