+++ src/find_of_Q_and_R.cpp
@@ -10,7 +10,7 @@
 int leadII_V::finding_of_Q(const int& peak){
 
  string operation  = "Q";
- vector <double> bufer;
+ vector <float> bufer;
 int ind,ind2,ind_vn;  
 ind =  peak - static_cast<int>(Fs*QRS.height);
 ind2 = peak;
@@ -20,7 +20,7 @@ set_indices(ind_vn,ind,ind2,count_iter,mem,mem_sdvig);
 if (peak - Fs*(QRS.height/2)>0 && ind>0 && ind2>ind){
 	 
       copy(begin(*ptr_signal)+ind,begin(*ptr_signal)+ind2 ,back_inserter(bufer));
-	  for_each(bufer.begin(), bufer.end(), [](double &n) { n = n*(-1); });
+	  for_each(bufer.begin(), bufer.end(), [](float &n) { n = n*(-1); });
 	 auto wave_peak =  max_element(bufer.begin(), bufer.end());
      Q.start = start_of_peak(bufer,ind_vn,operation);
 	 Q.peak =  distance(bufer.begin(), wave_peak) + ind_vn;
